backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
google-generativeai
hindsight-client
httpx
python-multipart
reportlab
